requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.11.3
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
annotated-types=0.7.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=4.6.2.post1=pypi_0
aom=3.6.1=h59595ed_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
asttokens=2.4.1=pypi_0
async-lru=2.0.4=pypi_0
attrs=24.2.0=pypi_0
babel=2.16.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
blas=2.116=mkl
blas-devel=3.9.0=16_linux64_mkl
bleach=6.2.0=pypi_0
brotli-python=1.1.0=py310hf71b8c6_2
bzip2=1.0.8=h4bc722e_7
ca-certificates=2024.12.14=hbcca054_0
certifi=2024.12.14=pyhd8ed1ab_0
cffi=1.17.1=py310h8deb56e_0
charset-normalizer=3.4.0=pyhd8ed1ab_0
click=8.1.7=pypi_0
comm=0.2.2=pypi_0
contourpy=1.3.1=pypi_0
cpython=3.10.15=py310hd8ed1ab_2
cuda-cudart=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.6.77=0
cuda-runtime=12.1.0=0
cuda-version=12.6=3
cycler=0.12.1=pypi_0
debugpy=1.8.9=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
dlib=19.24.6=pypi_0
docker-pycreds=0.4.0=pypi_0
eva-decord=0.6.1=pypi_0
exceptiongroup=1.2.2=pypi_0
executing=2.1.0=pypi_0
fastjsonschema=2.20.0=pypi_0
ffmpeg=4.4.2=gpl_hdf48244_113
filelock=3.16.1=pyhd8ed1ab_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=h77eed37_3
fontconfig=2.15.0=h7e30c49_1
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.55.0=pypi_0
fqdn=1.5.1=pypi_0
freetype=2.12.1=h267a509_2
fsspec=2024.10.0=pypi_0
gettext=0.22.5=he02047a_3
gettext-tools=0.22.5=he02047a_3
giflib=5.2.2=hd590300_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
gmp=6.3.0=hac33072_2
gmpy2=2.1.5=py310he8512ff_2
gnutls=3.7.9=hb077bed_0
h11=0.14.0=pypi_0
h2=4.1.0=pyhd8ed1ab_0
hpack=4.0.0=pyh9f0ad1d_0
httpcore=1.0.7=pypi_0
httpx=0.27.2=pypi_0
huggingface-hub=0.26.2=pypi_0
hydra-core=1.3.2=pypi_0
hyperframe=6.0.1=pyhd8ed1ab_0
idna=3.10=pyhd8ed1ab_0
imageio=2.36.1=pyh12aca89_1
imageio-ffmpeg=0.6.0=pypi_0
iopath=0.1.10=pypi_0
ipykernel=6.29.5=pypi_0
ipython=8.29.0=pypi_0
ipywidgets=8.1.5=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.19.2=pypi_0
jinja2=3.1.4=pyhd8ed1ab_0
json5=0.9.28=pypi_0
jsonpointer=3.0.0=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2024.10.1=pypi_0
jupyter=1.1.1=pypi_0
jupyter-client=8.6.3=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-core=5.7.2=pypi_0
jupyter-events=0.10.0=pypi_0
jupyter-lsp=2.2.5=pypi_0
jupyter-server=2.14.2=pypi_0
jupyter-server-terminals=0.5.3=pypi_0
jupyterlab=4.2.6=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.27.3=pypi_0
jupyterlab-widgets=3.0.13=pypi_0
kiwisolver=1.4.7=pypi_0
kornia=0.7.4=pypi_0
kornia-rs=0.1.7=pypi_0
lame=3.100=h166bdaf_1003
lcms2=2.16=hb7c19ff_0
ld_impl_linux-64=2.43=h712a8e2_2
lerc=4.0.0=h27087fc_0
libasprintf=0.22.5=he8f35ee_3
libasprintf-devel=0.22.5=he8f35ee_3
libblas=3.9.0=16_linux64_mkl
libcblas=3.9.0=16_linux64_mkl
libcublas=12.1.0.26=0
libcufft=11.0.2.4=0
libcufile=1.11.1.6=0
libcurand=10.3.7.77=0
libcusolver=11.4.4.55=0
libcusparse=12.0.2.55=0
libdeflate=1.22=hb9d3cd8_0
libdrm=2.4.123=hb9d3cd8_0
libegl=1.7.0=ha4b6fd6_2
libexpat=2.6.4=h5888daf_0
libffi=3.4.2=h7f98852_5
libgcc=14.2.0=h77fa898_1
libgcc-ng=14.2.0=h69a702a_1
libgettextpo=0.22.5=he02047a_3
libgettextpo-devel=0.22.5=he02047a_3
libgfortran=14.2.0=h69a702a_1
libgfortran-ng=14.2.0=h69a702a_1
libgfortran5=14.2.0=hd5240d6_1
libgl=1.7.0=ha4b6fd6_2
libglvnd=1.7.0=ha4b6fd6_2
libglx=1.7.0=ha4b6fd6_2
libgomp=14.2.0=h77fa898_1
libhwloc=2.11.1=default_hecaa2ac_1000
libiconv=1.17=hd590300_2
libidn2=2.3.7=hd590300_0
libjpeg-turbo=3.0.0=hd590300_1
liblapack=3.9.0=16_linux64_mkl
liblapacke=3.9.0=16_linux64_mkl
libnpp=12.0.2.50=0
libnsl=2.0.1=hd590300_0
libnvjitlink=12.1.105=0
libnvjpeg=12.1.1.14=0
libpciaccess=0.18=hd590300_0
libpng=1.6.44=hadc24fc_0
libsqlite=3.47.0=hadc24fc_1
libstdcxx=14.2.0=hc0a3c3a_1
libstdcxx-ng=14.2.0=h4852527_1
libtasn1=4.19.0=h166bdaf_0
libtiff=4.7.0=he137b08_1
libunistring=0.9.10=h7f98852_0
libuuid=2.38.1=h0b41bf4_0
libva=2.22.0=h8a09558_1
libvpx=1.13.1=h59595ed_0
libwebp=1.4.0=h2c329e2_0
libwebp-base=1.4.0=hd590300_0
libxcb=1.17.0=h8a09558_0
libxcrypt=4.4.36=hd590300_1
libxml2=2.13.5=h064dc61_0
libzlib=1.3.1=hb9d3cd8_2
llvm-openmp=15.0.7=h0cdce71_0
loguru=0.7.3=pypi_0
markupsafe=3.0.2=py310h89163eb_0
matplotlib=3.9.2=pypi_0
matplotlib-inline=0.1.7=pypi_0
mistune=3.0.2=pypi_0
mkl=2022.1.0=h84fe81f_915
mkl-devel=2022.1.0=ha770c72_916
mkl-include=2022.1.0=h84fe81f_915
mpc=1.3.1=h24ddda3_1
mpfr=4.2.1=h90cbb55_3
mpmath=1.3.0=pyhd8ed1ab_0
natsort=8.4.0=pypi_0
nbclient=0.10.0=pypi_0
nbconvert=7.16.4=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.5=he02047a_1
nest-asyncio=1.6.0=pypi_0
nettle=3.9.1=h7ab15ed_0
networkx=3.4.2=pyh267e887_2
notebook=7.2.2=pypi_0
notebook-shim=0.2.4=pypi_0
numpy=2.1.3=py310hd6e36ab_0
omegaconf=2.3.0=pypi_0
opencv-python=4.10.0.84=pypi_0
openexr=3.4.1=pypi_0
openh264=2.3.1=hcb278e6_2
openjpeg=2.5.2=h488ebb8_0
openssl=3.4.0=h7b32b05_1
overrides=7.7.0=pypi_0
p11-kit=0.24.1=hc5aa10d_0
packaging=24.2=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.4=pypi_0
pexpect=4.9.0=pypi_0
pillow=11.0.0=py310hfeaa1f3_0
pillow-avif-plugin=1.5.2=pypi_0
pip=24.3.1=pyh8b19718_0
platformdirs=4.3.6=pypi_0
plyfile=1.1=pyhd8ed1ab_0
portalocker=3.0.0=pypi_0
prometheus-client=0.21.0=pypi_0
prompt-toolkit=3.0.48=pypi_0
protobuf=5.29.1=pypi_0
psutil=6.1.0=pypi_0
pthread-stubs=0.4=hb9d3cd8_1002
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.3=pypi_0
pycparser=2.22=pyhd8ed1ab_0
pydantic=2.10.3=pypi_0
pydantic-core=2.27.1=pypi_0
pygments=2.18.0=pypi_0
pyparsing=3.2.0=pypi_0
pysocks=1.7.1=pyha2e5f31_6
python=3.10.15=h4a871b0_2_cpython
python-dateutil=2.9.0.post0=pypi_0
python-json-logger=2.0.7=pypi_0
python_abi=3.10=5_cp310
pytorch=2.5.1=py3.10_cuda12.1_cudnn9.1.0_0
pytorch-cuda=12.1=ha16c6d3_6
pytorch-mutex=1.0=cuda
pyyaml=6.0.2=py310ha75aee5_1
pyzmq=26.2.0=pypi_0
readline=8.2=h8228510_1
referencing=0.35.1=pypi_0
regex=2024.11.6=pypi_0
requests=2.32.3=pyhd8ed1ab_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rpds-py=0.21.0=pypi_0
safetensors=0.4.5=pypi_0
sam-2=1.0=pypi_0
scipy=1.14.1=py310hfcf56fc_1
send2trash=1.8.3=pypi_0
sentry-sdk=2.19.2=pypi_0
setproctitle=1.3.4=pypi_0
setuptools=75.6.0=pyhff2d567_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.6=pypi_0
stack-data=0.6.3=pypi_0
svt-av1=1.4.1=hcb278e6_0
sympy=1.13.1=pypi_0
tbb=2021.13.0=h84d6215_0
tensorboardx=2.6.2.2=pypi_0
termcolor=2.5.0=pypi_0
terminado=0.18.1=pypi_0
timm=1.0.11=pypi_0
tinycss2=1.4.0=pypi_0
tk=8.6.13=noxft_h4845f30_101
tokenizers=0.21.0=pypi_0
tomli=2.1.0=pypi_0
torchaudio=2.5.1=py310_cu121
torchtriton=3.1.0=py310
torchvision=0.20.1=py310_cu121
tornado=6.4.2=pypi_0
tqdm=4.67.0=pypi_0
traitlets=5.14.3=pypi_0
transformers=4.47.0=pypi_0
trimesh=4.5.3=pyhd8ed1ab_1
types-python-dateutil=2.9.0.20241003=pypi_0
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024b=hc8b5060_0
uri-template=1.3.0=pypi_0
urllib3=2.2.3=pyhd8ed1ab_0
wandb=0.19.0=pypi_0
wayland=1.23.1=h3e06ad9_0
wayland-protocols=1.37=hd8ed1ab_0
wcwidth=0.2.13=pypi_0
webcolors=24.11.1=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.8.0=pypi_0
wheel=0.45.1=pyhd8ed1ab_0
widgetsnbextension=4.0.13=pypi_0
x264=1!164.3095=h166bdaf_2
x265=3.5=h924138e_3
xorg-libx11=1.8.10=h4f16b4b_0
xorg-libxau=1.0.11=hb9d3cd8_1
xorg-libxdmcp=1.1.5=hb9d3cd8_0
xorg-libxext=1.3.6=hb9d3cd8_0
xorg-libxfixes=6.0.1=hb9d3cd8_0
xorg-xorgproto=2024.1=hb9d3cd8_1
xz=5.2.6=h166bdaf_0
yacs=0.1.8=pypi_0
yaml=0.2.5=h7f98852_2
zstandard=0.23.0=py310ha39cb0e_1
zstd=1.5.6=ha6fb4c9_0
